Add restartretries, stop and start time functionality

+++ pysrcs/processes.py
@@ -6,6 +6,7 @@ import os
 import signal
 import time
 import threading
+import sys
 
 import output
 import menuloop
@@ -13,18 +14,16 @@ import menuloop
 def stop_time(saved_pid):
 	for program in menuloop.globProgramList:
 		for pid in program.pidList:
-			if saved_pid == pid:
+			if saved_pid == pid[0]:
 				if pid[1] == "Stopping":
 						pid[1] = "Stopped"
-						# break
 
 def start_time(saved_pid):
 	for program in menuloop.globProgramList:
 		for pid in program.pidList:
-			if saved_pid == pid:
+			if saved_pid == pid[0]:
 				if pid[1] == "Starting":
 						pid[1] = "Running"
-						# break
 
 def stop_program(programList):
 	"""this function stops a program with the desired signal"""
@@ -46,10 +45,17 @@ def stop_program(programList):
 				for pid in program.pidList:
 					if pid[1] != "Stopped" and pid[1] != "Finished" and pid[1] != "Stopping":
 							os.kill(pid[0].pid, s)
-							pid[1] = "Stopping"
-							timer = threading.Timer(program.stoptime, stop_time, [pid[0]])
-							timer.start()
-				program.state = "Stopping"
+							if program.stoptime > 0:
+								pid[1] = "Stopping"
+								timer = threading.Timer(program.stoptime, stop_time, [pid[0]])
+								timer.daemon = True
+								timer.start()
+							else:
+								pid[1] = "Stopped"
+				if program.stoptime > 0:
+					program.state = "Stopping"
+				else:
+					program.state = "Stopped"
 			else:
 				print(output.bcolors.FAIL + "Program " + program.name + " was already stopped/finished/killed or hadn't started!" + output.bcolors.ENDC)
 	return 0
@@ -58,14 +64,16 @@ def restart_program(programList):
 	"""this function restarts a program"""
 	for program in programList:
 		if program.selected == 1:
-			for pid in program.pidList:
-				if pid[1] != "Stopped" and pid[1] != "Finished" and pid[1] != "Stopping":
-					os.kill(pid[0].pid, signal.SIGKILL)
-			program.state = "Not started"
-			program.started = False
-			program.pidList = []
-			if program.autostart == 1:
-				program.state = "Starting"
+			if program.state != "Not started":
+				for pid in program.pidList:
+					if pid[1] != "Stopped" and pid[1] != "Finished" and pid[1] != "Stopping":
+						os.kill(pid[0].pid, signal.SIGKILL)
+				program.pidList = []
+				program.started = True
+				if program.starttime > 0:
+					program.state = "Starting"
+				else:
+					program.state = "Running"
 				envcopy = os.environ.copy()
 				if program.env != "None" and isinstance(program.env, list):
 					for envitem in program.env:
@@ -87,23 +95,38 @@ def restart_program(programList):
 							errpath = program.stderr
 				else:
 					errpath = "/dev/null"
-				if program.autostart == True:
-					program.started = True
-					cmdList = program.cmd.split()
-					instances = program.cmdammount
-					while instances > 0:
+				cmdList = program.cmd.split()
+				instances = program.cmdammount
+				while instances > 0:
+					alarm = 0
+					retries = program.restartretries
+					while retries > 0:
 						try:
 							with open(outpath, "wb", 0) as out, open(errpath, "wb", 0) as err:
-								proc = subprocess.Popen(cmdList, stdout=out, stderr=err, env=envcopy, preexec_fn=initchildproc(program))
-								timer = threading.Timer(program.starttime, start_time, [proc])
-								timer.start()
+								proc = subprocess.Popen(cmdList, stdout=out, stderr=err, env=envcopy, preexec_fn=execution.initchildproc(program))
+								break
 						except:
-							print("Could not run the subprocess for", program.name,
-							"skipping this execution")
-							break
+							if retries > 0:
+								print("Could not run the subprocess for", program.name, end='')
+								print(f". retries left: {retries}")
+								retries -= 1
+								if retries == 0:
+									alarm = 1
+									print("Could not run the subprocess for", program.name,
+									"skipping this execution")
+								continue
+					if alarm == 1:
+						break
+					if program.starttime > 0:
 						program.pidList.append([proc, "Starting", None])
-						instances -= 1
-					program.state = "Starting"
+						timer = threading.Timer(program.starttime, start_time, [proc])
+						timer.daemon = True
+						timer.start()
+					else:
+						program.pidList.append([proc, "Running", None])
+					instances -= 1
+			else:
+				print(f"The program: {program.name} had never been started, so restarting is not possible!", file=sys.stderr)
 	return programList
 
 def start_program(programList):
@@ -136,21 +159,40 @@ def start_program(programList):
 				cmdList = program.cmd.split()
 				instances = program.cmdammount
 				while instances > 0:
-					try:
-						with open(outpath, "wb", 0) as out, open(errpath, "wb", 0) as err:
-							proc = subprocess.Popen(cmdList, stdout=out, 
-													stderr=err, 
-													env=envcopy,
-													preexec_fn=execution.initchildproc(program))
-							timer = threading.Timer(program.starttime, start_time, [proc])
-							timer.start()
-					except:
-						print("Could not run the subprocess for", program.name,
-						"skipping this execution")
+					alarm = 0
+					retries = program.restartretries
+					while retries:
+						try:
+							with open(outpath, "wb", 0) as out, open(errpath, "wb", 0) as err:
+								proc = subprocess.Popen(cmdList, stdout=out, 
+														stderr=err, 
+														env=envcopy,
+														preexec_fn=execution.initchildproc(program))
+								break
+						except:
+							if retries > 0:
+								print("Could not run the subprocess for", program.name, end='')
+								print(f". retries left: {retries}")
+								retries -= 1
+								if retries == 0:
+									alarm = 1
+									print("Could not run the subprocess for", program.name,
+									"skipping this execution")
+								continue
+					if alarm == 1:
 						break
-					program.pidList.append([proc, "Starting", None])
+					if program.starttime > 0:
+						program.pidList.append([proc, "Starting", None])
+						timer = threading.Timer(program.starttime, start_time, [proc])
+						timer.daemon = True
+						timer.start()
+					else:
+						program.pidList.append([proc, "Running", None])
 					instances -= 1
-				program.state = "Starting"
+				if program.starttime > 0:
+					program.state = "Starting"
+				else:
+					program.state = "Running"
 			else:
 				print(output.bcolors.FAIL + "Program " + program.name + " was already started/stopped!" + output.bcolors.ENDC)
 	return 0
